Create Building a simple decision tree
Building a simple decision tree
The loans dataset contains 11,312 randomly-selected people who applied for and later received loans from Lending Club, a US-based peer-to-peer lending company.

You will use a decision tree to try to learn patterns in the outcome of these loans (either repaid or default) based on the requested loan amount and credit score at the time of application.

Then, see how the tree's predictions differ for an applicant with good credit versus one with bad credit.

The dataset loans has been loaded for you.

Instructions

Load the rpart package.
Fit a decision tree model with the function rpart().
Supply the R formula that specifies outcome as a function of loan_amount and credit_score as the first argument.
Leave the control argument alone for now. (You'll learn more about that later!)
Use predict() with the resulting loan model to predict the outcome for the good_credit applicant. Use the type argument to predict the "class" of the outcome.
Do the same for the bad_credit applicant.