Create Building a random forest model
Building a random forest model
In spite of the fact that a forest can contain hundreds of trees, growing a decision tree forest is perhaps even easier than creating a single highly-tuned tree.

Using the randomForest package, build a random forest and see how it compares to the single trees you built previously.

Keep in mind that due to the random nature of the forest, the results may vary slightly each time you create the forest.

Instructions

Load the randomForest package.
Build a random forest model using all of the loan application variables. The randomForest function also uses the formula interface.
Compute the accuracy of the random forest model to compare to the original tree's accuracy of 57.6% using predict() and mean().